refactor: change CheckAnswerRequest logiс ss-031
